requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.3
notion-client>=2.0.0
python-dotenv>=1.0.0
